refactor: move Record related schemas to his own file and add set dataset_id attribute always to Record schema
