+++ src/argilla/server/contexts/datasets.py
@@ -56,18 +56,25 @@ from argilla.server.schemas.v1.datasets import (
     FieldCreate,
     MetadataPropertyCreate,
     QuestionCreate,
-    RecordCreate,
-    RecordIncludeParam,
-    RecordsCreate,
-    RecordUpdateWithId,
-    ResponseValueCreate,
 )
 from argilla.server.schemas.v1.datasets import (
     VectorSettings as VectorSettingsSchema,
 )
 from argilla.server.schemas.v1.metadata_properties import MetadataPropertyUpdate
-from argilla.server.schemas.v1.records import ResponseCreate
-from argilla.server.schemas.v1.responses import ResponseUpdate, ResponseUpsert, ResponseValueUpdate
+from argilla.server.schemas.v1.records import (
+    RecordCreate,
+    RecordIncludeParam,
+    RecordsCreate,
+    RecordsUpdate,
+    RecordUpdateWithId,
+)
+from argilla.server.schemas.v1.responses import (
+    ResponseCreate,
+    ResponseUpdate,
+    ResponseUpsert,
+    ResponseValueCreate,
+    ResponseValueUpdate,
+)
 from argilla.server.schemas.v1.vectors import Vector as VectorSchema
 from argilla.server.search_engine import SearchEngine
 from argilla.server.security.model import User
@@ -77,7 +84,6 @@ if TYPE_CHECKING:
 
     from argilla.server.schemas.v1.datasets import (
         DatasetUpdate,
-        RecordsUpdate,
         VectorSettingsCreate,
     )
     from argilla.server.schemas.v1.fields import FieldUpdate
