+++ tests/unit/server/api/v1/datasets/test_search_dataset_records.py
@@ -128,6 +128,7 @@ class TestSearchDatasetRecords:
                                 "id": str(response_a.id),
                                 "status": "submitted",
                                 "values": {"input_ok": {"value": "yes"}},
+                                "record_id": str(record_a.id),
                                 "user_id": str(response_a.user_id),
                                 "inserted_at": response_a.inserted_at.isoformat(),
                                 "updated_at": response_a.updated_at.isoformat(),
@@ -136,12 +137,14 @@ class TestSearchDatasetRecords:
                                 "id": str(response_b.id),
                                 "status": "submitted",
                                 "values": {"input_ok": {"value": "no"}},
+                                "record_id": str(record_a.id),
                                 "user_id": str(response_b.user_id),
                                 "inserted_at": response_b.inserted_at.isoformat(),
                                 "updated_at": response_b.updated_at.isoformat(),
                             },
                         ],
                         "external_id": record_a.external_id,
+                        "dataset_id": str(record_a.dataset_id),
                         "inserted_at": record_a.inserted_at.isoformat(),
                         "updated_at": record_a.updated_at.isoformat(),
                     },
@@ -160,12 +163,14 @@ class TestSearchDatasetRecords:
                                 "id": str(response_c.id),
                                 "status": "submitted",
                                 "values": {"input_ok": {"value": "yes"}},
+                                "record_id": str(record_b.id),
                                 "user_id": str(response_c.user_id),
                                 "inserted_at": response_c.inserted_at.isoformat(),
                                 "updated_at": response_c.updated_at.isoformat(),
                             },
                         ],
                         "external_id": record_b.external_id,
+                        "dataset_id": str(record_b.dataset_id),
                         "inserted_at": record_b.inserted_at.isoformat(),
                         "updated_at": record_b.updated_at.isoformat(),
                     },
