+++ tests/unit/server/api/v1/test_datasets.py
@@ -47,14 +47,9 @@ from argilla.server.schemas.v1.datasets import (
     FIELD_CREATE_TITLE_MAX_LENGTH,
     METADATA_PROPERTY_CREATE_NAME_MAX_LENGTH,
     METADATA_PROPERTY_CREATE_TITLE_MAX_LENGTH,
-    QUESTION_CREATE_DESCRIPTION_MAX_LENGTH,
-    QUESTION_CREATE_NAME_MAX_LENGTH,
-    QUESTION_CREATE_TITLE_MAX_LENGTH,
     RANKING_OPTIONS_MAX_ITEMS,
     RATING_OPTIONS_MAX_ITEMS,
     RATING_OPTIONS_MIN_ITEMS,
-    RECORDS_CREATE_MAX_ITEMS,
-    RECORDS_CREATE_MIN_ITEMS,
     TERMS_METADATA_PROPERTY_VALUES_MAX_ITEMS,
     VALUE_TEXT_OPTION_DESCRIPTION_MAX_LENGTH,
     VALUE_TEXT_OPTION_TEXT_MAX_LENGTH,
@@ -62,6 +57,12 @@ from argilla.server.schemas.v1.datasets import (
     VECTOR_SETTINGS_CREATE_NAME_MAX_LENGTH,
     VECTOR_SETTINGS_CREATE_TITLE_MAX_LENGTH,
 )
+from argilla.server.schemas.v1.questions import (
+    QUESTION_CREATE_DESCRIPTION_MAX_LENGTH,
+    QUESTION_CREATE_NAME_MAX_LENGTH,
+    QUESTION_CREATE_TITLE_MAX_LENGTH,
+)
+from argilla.server.schemas.v1.records import RECORDS_CREATE_MAX_ITEMS, RECORDS_CREATE_MIN_ITEMS
 from argilla.server.search_engine import (
     FloatMetadataFilter,
     IntegerMetadataFilter,
@@ -3982,6 +3983,7 @@ class TestSuiteDatasets:
                         "fields": {"input": "input_a", "output": "output_a"},
                         "metadata": None,
                         "external_id": records[0].external_id,
+                        "dataset_id": str(records[0].dataset_id),
                         "inserted_at": records[0].inserted_at.isoformat(),
                         "updated_at": records[0].updated_at.isoformat(),
                     },
@@ -3993,6 +3995,7 @@ class TestSuiteDatasets:
                         "fields": {"input": "input_b", "output": "output_b"},
                         "metadata": {"unit": "test"},
                         "external_id": records[1].external_id,
+                        "dataset_id": str(records[1].dataset_id),
                         "inserted_at": records[1].inserted_at.isoformat(),
                         "updated_at": records[1].updated_at.isoformat(),
                     },
@@ -4335,6 +4338,7 @@ class TestSuiteDatasets:
                         },
                         "metadata": None,
                         "external_id": records[0].external_id,
+                        "dataset_id": str(records[0].dataset_id),
                         "inserted_at": records[0].inserted_at.isoformat(),
                         "updated_at": records[0].updated_at.isoformat(),
                     },
@@ -4349,6 +4353,7 @@ class TestSuiteDatasets:
                         },
                         "metadata": {"unit": "test"},
                         "external_id": records[1].external_id,
+                        "dataset_id": str(records[1].dataset_id),
                         "inserted_at": records[1].inserted_at.isoformat(),
                         "updated_at": records[1].updated_at.isoformat(),
                     },
@@ -4367,6 +4372,7 @@ class TestSuiteDatasets:
                     "id": str(first_owner_response.id),
                     "values": None,
                     "status": "discarded",
+                    "record_id": str(records[0].id),
                     "user_id": str(owner.id),
                     "inserted_at": first_owner_response.inserted_at.isoformat(),
                     "updated_at": first_owner_response.updated_at.isoformat(),
@@ -4380,6 +4386,7 @@ class TestSuiteDatasets:
                         "output_ok": {"value": "no"},
                     },
                     "status": "submitted",
+                    "record_id": str(records[1].id),
                     "user_id": str(owner.id),
                     "inserted_at": second_owner_response.inserted_at.isoformat(),
                     "updated_at": second_owner_response.updated_at.isoformat(),
@@ -4481,6 +4488,7 @@ class TestSuiteDatasets:
                             "vector-a": [1.0, 2.0, 3.0],
                             "vector-b": [4.0, 5.0],
                         },
+                        "dataset_id": str(record_a.dataset_id),
                         "inserted_at": record_a.inserted_at.isoformat(),
                         "updated_at": record_a.updated_at.isoformat(),
                     },
@@ -4495,6 +4503,7 @@ class TestSuiteDatasets:
                         "vectors": {
                             "vector-b": [1.0, 2.0],
                         },
+                        "dataset_id": str(record_b.dataset_id),
                         "inserted_at": record_b.inserted_at.isoformat(),
                         "updated_at": record_b.updated_at.isoformat(),
                     },
@@ -4507,6 +4516,7 @@ class TestSuiteDatasets:
                         "metadata": None,
                         "external_id": record_c.external_id,
                         "vectors": {},
+                        "dataset_id": str(record_c.dataset_id),
                         "inserted_at": record_c.inserted_at.isoformat(),
                         "updated_at": record_c.updated_at.isoformat(),
                     },
@@ -4572,6 +4582,7 @@ class TestSuiteDatasets:
                             "vector-a": [1.0, 2.0, 3.0],
                             "vector-b": [4.0, 5.0],
                         },
+                        "dataset_id": str(record_a.dataset_id),
                         "inserted_at": record_a.inserted_at.isoformat(),
                         "updated_at": record_a.updated_at.isoformat(),
                     },
@@ -4586,6 +4597,7 @@ class TestSuiteDatasets:
                         "vectors": {
                             "vector-b": [1.0, 2.0],
                         },
+                        "dataset_id": str(record_b.dataset_id),
                         "inserted_at": record_b.inserted_at.isoformat(),
                         "updated_at": record_b.updated_at.isoformat(),
                     },
@@ -4598,6 +4610,7 @@ class TestSuiteDatasets:
                         "metadata": None,
                         "external_id": record_c.external_id,
                         "vectors": {},
+                        "dataset_id": str(record_c.dataset_id),
                         "inserted_at": record_c.inserted_at.isoformat(),
                         "updated_at": record_c.updated_at.isoformat(),
                     },
