+++ tests/unit/server/api/v1/test_records.py
@@ -98,6 +98,7 @@ class TestSuiteRecords:
             "responses": None,
             "suggestions": [],
             "vectors": None,
+            "dataset_id": str(dataset.id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -213,6 +214,7 @@ class TestSuiteRecords:
                 vector_settings_1.name: [2.0, 2.0, 2.0, 2.0, 2.0],
                 vector_settings_2.name: [3.0, 3.0, 3.0, 3.0, 3.0],
             },
+            "dataset_id": str(dataset.id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -245,6 +247,7 @@ class TestSuiteRecords:
             "responses": [],
             "suggestions": [],
             "vectors": {},
+            "dataset_id": str(dataset.id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -271,6 +274,7 @@ class TestSuiteRecords:
             "responses": None,
             "suggestions": [],
             "vectors": {},
+            "dataset_id": str(dataset.id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -304,6 +308,7 @@ class TestSuiteRecords:
             "responses": [],
             "suggestions": [],
             "vectors": {},
+            "dataset_id": str(dataset.id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -330,6 +335,7 @@ class TestSuiteRecords:
             "responses": None,
             "suggestions": [],
             "vectors": {},
+            "dataset_id": str(record.dataset_id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
@@ -567,6 +573,7 @@ class TestSuiteRecords:
             "id": str(UUID(response_body["id"])),
             "values": responses["values"],
             "status": response_status,
+            "record_id": str(record.id),
             "user_id": str(owner.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -622,6 +629,7 @@ class TestSuiteRecords:
             "id": str(UUID(response_body["id"])),
             "values": responses["values"],
             "status": response_status,
+            "record_id": str(record.id),
             "user_id": str(owner.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -672,6 +680,7 @@ class TestSuiteRecords:
             "id": str(UUID(response_body["id"])),
             "values": {question.name: {"value": response_value}},
             "status": ResponseStatus.submitted,
+            "record_id": str(record.id),
             "user_id": str(owner.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -727,6 +736,7 @@ class TestSuiteRecords:
             "id": str(UUID(response_body["id"])),
             "values": {question.name: {"value": response_value}},
             "status": response_status.value,
+            "record_id": str(record.id),
             "user_id": str(owner.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -982,6 +992,7 @@ class TestSuiteRecords:
                 "output_ok": {"value": "yes"},
             },
             "status": status,
+            "record_id": str(record.id),
             "user_id": str(owner.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -1018,6 +1029,7 @@ class TestSuiteRecords:
                 "id": str(UUID(response_body["id"])),
                 "values": None,
                 "status": status,
+                "record_id": str(record.id),
                 "user_id": str(owner.id),
                 "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
                 "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -1069,6 +1081,7 @@ class TestSuiteRecords:
                 "output_ok": {"value": "yes"},
             },
             "status": "submitted",
+            "record_id": str(record.id),
             "user_id": str(user.id),
             "inserted_at": datetime.fromisoformat(response_body["inserted_at"]).isoformat(),
             "updated_at": datetime.fromisoformat(response_body["updated_at"]).isoformat(),
@@ -1342,6 +1355,7 @@ class TestSuiteRecords:
             "fields": record.fields,
             "metadata": None,
             "external_id": record.external_id,
+            "dataset_id": str(record.dataset_id),
             "inserted_at": record.inserted_at.isoformat(),
             "updated_at": record.updated_at.isoformat(),
         }
